+++ src/main/java/io/github/astrapi69/io/file/FileConstants.java
@@ -33,26 +33,45 @@ package io.github.astrapi69.io.file;
 public abstract class FileConstants
 {
 
-	/** Constant for the slash. current value:"\" */
+	/**
+	 * Constant for the slash. current value:"\"
+	 */
 	public static final String BACKSLASH = "\\";
 
-	/** Constant for the blocksize. current value:8192 */
+	/**
+	 * Constant for the blocksize. current value:8192
+	 */
 	public static final int BLOCKSIZE = 8192;
 
-	/** Constant for the dot. current value:"." */
+	/**
+	 * Constant for the dot. current value:"."
+	 */
 	public static final String DOT = ".";
 
-	/** Constant for the dot. current value:":" */
+	/**
+	 * Constant for the dot. current value:":"
+	 */
 	public static final String DOUBLEDOT = ":";
-	/** Constant for the kilobyte. current value:1024 */
+
+	/**
+	 * Constant for the kilobyte. current value:1024
+	 */
 	public static final int KILOBYTE = 1024;
-	/** Constant for the slash. current value:"/" */
+
+	/**
+	 * Constant for the slash. current value:"/"
+	 */
 	public static final String SLASH = "/";
+
 	/**
 	 * String array with a few extensions from zip-files
 	 */
-	public static final String[] ZIP_EXTENSIONS = { ".zip", ".jar", ".war", ".ear", ".tar", ".rar",
-			".7z", ".bz2", ".gz" };
+	public static final String[] ZIP_EXTENSIONS = { FileExtension.ZIP.getExtension(),
+			FileExtension.JAR.getExtension(), FileExtension.WAR.getExtension(),
+			FileExtension.EAR.getExtension(), FileExtension.TAR.getExtension(),
+			FileExtension.RAR.getExtension(), FileExtension.SEVEN_ZIP.getExtension(),
+			FileExtension.BZ2.getExtension(), FileExtension.GZ.getExtension() };
+
 	/**
 	 * String array with the invalid characters in filenames
 	 */