+++ src/main/java/io/github/astrapi69/io/file/FileExtension.java
@@ -24,6 +24,9 @@
  */
 package io.github.astrapi69.io.file;
 
+import java.io.File;
+import java.util.Objects;
+
 /**
  * The enum {@link FileExtension} for file extensions
  */
@@ -48,6 +51,33 @@ public enum FileExtension
 	/** The file extension for java files */
 	JAVA(".java"),
 
+	/** The file extension for java archive files */
+	JAR(".jar"),
+
+	/** The file extension for archive files created by tar unix utility program */
+	TAR(".tar"),
+
+	/** The file extension for archive files created with RAR compression */
+	RAR(".rar"),
+
+	/** The file extension for java web archive files */
+	WAR(".war"),
+
+	/** The file extension for java enterprise archive files */
+	EAR(".ear"),
+
+	/** The file extension for archive files created with zip compression */
+	ZIP(".zip"),
+
+	/** The file extension for archive files created with bzip2 unix compression program */
+	BZ2(".bz2"),
+
+	/** The file extension for archive files created with gzip compression algorithm */
+	GZ(".gz"),
+
+	/** The file extension for archive files created with 7z compression */
+	SEVEN_ZIP(".7z"),
+
 	/** The file extension for properties files */
 	PROPERTIES(".properties"),
 
@@ -69,6 +99,40 @@ public enum FileExtension
 	FileExtension(final String extension)
 	{
 		this.extension = extension;
+	}
+
+	/**
+	 * Checks if the given file is from the given {@link FileExtension} object
+	 *
+	 * @param file
+	 *            the file to check
+	 * @param fileExtension
+	 *            the file extension
+	 * @return true if the given file is from the given {@link FileExtension} object otherwise false
+	 */
+	public static boolean is(File file, FileExtension fileExtension)
+	{
+		Objects.requireNonNull(file, "File must not be null");
+		Objects.requireNonNull(file, "FileExtension must not be null");
+		return is(file, fileExtension.getExtension());
+	}
+
+	/**
+	 * Checks if the given file is from the given file extension as {@link String} object
+	 *
+	 * @param file
+	 *            the file to check
+	 * @param fileExtension
+	 *            the file extension as {@link String} object
+	 * @return true if the given file is from the given {@link String} object otherwise false
+	 */
+	public static boolean is(File file, String fileExtension)
+	{
+		Objects.requireNonNull(file, "File must not be null");
+		Objects.requireNonNull(file, "String for the file extension must not be null");
+		return fileExtension.startsWith(".")
+			? FilenameExtensions.getFileExtension(file, true).equals(fileExtension)
+			: FilenameExtensions.getFileExtension(file, false).equals(fileExtension);
 	}
 
 	public String getExtension()