+++ src/main/java/io/github/astrapi69/io/file/FilenameExtensions.java
@@ -46,17 +46,59 @@ public final class FilenameExtensions
 	public static String getFilenamePrefix(final File file)
 	{
 		final String fileName = file.getAbsolutePath();
-		final int ext_index = fileName.lastIndexOf(".");
+		final int filenameEndIndex = fileName.lastIndexOf(".");
 		final String fileNamePrefix;
-		if (ext_index != -1)
+		if (filenameEndIndex != -1)
 		{
-			fileNamePrefix = fileName.substring(0, ext_index);
+			fileNamePrefix = fileName.substring(0, filenameEndIndex);
 		}
 		else
 		{
 			fileNamePrefix = fileName;
 		}
 		return fileNamePrefix;
+	}
+
+	/**
+	 * Gets the file extension or an empty String if no file extension exists or the given file
+	 * object is a directory
+	 *
+	 * @param file
+	 *            the file
+	 * @return 's the file extension or an empty String if no file extension exists or the given
+	 *         file object is a directory
+	 */
+	public static String getFileExtension(final File file)
+	{
+		return getFileExtension(file, false);
+	}
+
+	/**
+	 * Gets the file extension or an empty String if no file extension exists or the given file
+	 * object is a directory
+	 *
+	 * @param file
+	 *            the file
+	 * @param withDot
+	 *            the flag that indicates if the extension shell have a dot at the start
+	 * @return 's the file extension or an empty String if no file extension exists or the given
+	 *         file object is a directory
+	 */
+	public static String getFileExtension(final File file, boolean withDot)
+	{
+		String fileNameSuffix = "";
+		if (!file.isDirectory())
+		{
+			final String fileName = file.getName();
+			final int filenameEndIndex = fileName.lastIndexOf(".");
+
+			if (filenameEndIndex != -1)
+			{
+				int beginIndex = withDot ? filenameEndIndex : filenameEndIndex + 1;
+				fileNameSuffix = fileName.substring(beginIndex);
+			}
+		}
+		return fileNameSuffix;
 	}
 
 	/**
@@ -69,22 +111,17 @@ public final class FilenameExtensions
 	 */
 	public static String getFilenameSuffix(final File file)
 	{
+		String fileNameSuffix = null;
 		if (!file.isDirectory())
 		{
 			final String fileName = file.getAbsolutePath();
-			final int ext_index = fileName.lastIndexOf(".");
-			final String fileNameSuffix;
-			if (ext_index != -1)
+			final int filenameEndIndex = fileName.lastIndexOf(".");
+			if (filenameEndIndex != -1)
 			{
-				fileNameSuffix = fileName.substring(ext_index, fileName.length());
+				fileNameSuffix = fileName.substring(filenameEndIndex, fileName.length());
 			}
-			else
-			{
-				fileNameSuffix = null;
-			}
-			return fileNameSuffix;
 		}
-		return null;
+		return fileNameSuffix;
 	}
 
 	/**